Addressed PR comment - moved accounts for testing
